cppcheck-codequality and cppcheck-htmlreport added
cppcheck-htmlreport has been around for a while,
but it didn't work due to the lack of a python
runtime.

It is important that the size of the image remains
as small as possible, so a binary is created from
the cppcheck-codequality and cppcheck-htmlreport
program using pyinstaller, so the python runtime
environment is still not required.